GFDataLoadingVC is created. VC's with loading screen inherited GFDataLoadingVC
