+++ TakeHomeProject/Screens/FollowerListVC.swift
@@ -11,7 +11,7 @@ protocol FollowerListVCDelegate: AnyObject {
     func didRequestFollowers(for username: String)
 }
 
-class FollowerListVC: UIViewController {
+class FollowerListVC: GFDataLoadingVC {
     
     ///Enums are hashable by default
     enum Section {
